few things concerning packages
So, it looks like the packages are now updating properly, maybe - at least they are appearing in the package.json.

A couple of things to note, while playing with echarts and ngx-echarts (which I was unable to get working, though now thinking about it likely because I did not import it in the actual module I was using it) the versions have been updated but no idea if these actually work, the other updates don't appear to have had a negative impact so will leave them as they are for the moment, but if strange things start to happen shortly this is the culprit probably.