add logout feature, small tweaks
somehow managed to work this out and get it working, with a couple of lines of code. Either I am very smart, or just lucky this time. I would go with luck.

Also, added theme for ocean signal selectable, and also the other default themes, who knows someone might have a preference for one of the others...

Finally, added loading spinners for the support programs