Finally get the radio buttons to be hoz aligning
Started to lose the will to live here, something so simple sucked a good 3 hours - although bootstrap supports it, the nebular ui must be doing something else, found someone else who had the issue and sorted it out with a bit of custom css. It will do, the override is only specific to this page.